+++ src/css/header.css
@@ -83,7 +83,7 @@
     /* display: none; */
     display: block;
     position: absolute;
-    top: 30%;
+    top: 50%;
     left: 50%;
     /* transform: translateX(-50%); */
     transform: translateX(-50%) scaleY(0);
@@ -114,7 +114,7 @@
 }
 
 
-.custom-checkbox:checked ~ .custom-list {
+.custom-checkbox:checked~.custom-list {
     transform: translateX(-50%) scaleY(1);
     opacity: 1;
 
@@ -250,7 +250,8 @@
     color: #fafafa;
 }
 
-.order-mobile:hover, .order-mobile:focus {
+.order-mobile:hover,
+.order-mobile:focus {
     color: var(--aсcent-collor-hover);
     transition: var(--trns);
 }
@@ -271,7 +272,7 @@
         position: relative;
         width: max-content;
         margin: 0 auto;
-        padding-top: 10px;
+        padding-top: 13px;
     }
 
     .menu-list-item {
@@ -297,17 +298,18 @@
         color: #fafafa;
     }
 
-    .order-btn:hover, .order-btn:focus {
+    .order-btn:hover,
+    .order-btn:focus {
         color: var(--aсcent-collor-hover);
         transition: var(--trns);
     }
 
-    .change-theme{
-         width: 90px;
-        height: 45px; 
+    .change-theme {
+        width: 90px;
+        height: 45px;
         transform: translateX(0);
-    
-               
+
+
     }
 }
 
@@ -318,8 +320,8 @@
         gap: 14px;
     }
 
-        .change-theme {
-            /* width: 60px;
+    .change-theme {
+        /* width: 60px;
             height: 35px; */
-        }
+    }
 }
